refactor: split scrollmagic context per chunks
