+++ src/components/ridipaper4/RidiPaper4ScrollmagicContext.tsx
@@ -1,0 +1,3 @@
+import { createScrollmagicContext } from '@/utils/createScrollmagicContext/createScrollmagicContext';
+
+export const { useScrollmagicEffect, ScrollmagicProvider } = createScrollmagicContext();
