+++ src/components/teaser/TeaserScrollmagicContext.tsx
@@ -1,0 +1,3 @@
+import { createScrollmagicContext } from '@/utils/createScrollmagicContext/createScrollmagicContext';
+
+export const { useScrollmagicEffect, ScrollmagicProvider } = createScrollmagicContext();
