+++ src/pages/ridipaper.tsx
@@ -2,9 +2,8 @@ import { css } from 'astroturf';
 import React from 'react';
 
 import { graphql, useStaticQuery } from 'gatsby';
-import { Controller } from 'scrollmagic';
 
-import { ScrollmagicProvider } from '@/components/common/ScrollMagicContext';
+import { ScrollmagicProvider } from '@/components/ridipaper/RidiPaperScrollMagicContext';
 import Layout from '@/components/common/Layout';
 import SEO from '@/components/common/SEO';
 
@@ -42,7 +41,6 @@ const styles = css`
 `;
 
 const RidiPaperPage = () => {
-  const [controller, setController] = React.useState<Controller>();
   const data = useStaticQuery(graphql`
     {
       banner: file(relativePath: { eq: "images/meta/ridipaper/og.jpg" }) {
@@ -51,26 +49,9 @@ const RidiPaperPage = () => {
     }
   `);
 
-  React.useEffect(() => {
-    let controller: Controller;
-    let destroyed = false;
-    import('scrollmagic').then(({ Controller }) => {
-      if (destroyed) {
-        return;
-      }
-      controller = new Controller();
-      setController(controller);
-    });
-
-    return () => {
-      controller && controller.destroy();
-      destroyed = true;
-    };
-  }, []);
-
   return (
     <Layout className={styles.main}>
-      <ScrollmagicProvider value={controller}>
+      <ScrollmagicProvider>
         <SEO
           meta={[{ property: 'og:image', content: data.banner.publicURL }]}
         />
